feat(core): split motion logic into easing module
Extract easing functionality from builtin_motion into a new dedicated module to improve separation of concerns and prepare for future easing extensions.
